move peak_pipeline into core and rename it as node_pipeline. Change tests accordingly
