+++ src/spikeinterface/sortingcomponents/tests/test_waveforms/test_neural_network_denoiser.py
@@ -5,7 +5,7 @@ from spikeinterface.extractors import MEArecRecordingExtractor
 from spikeinterface import download_dataset
 
 
-from spikeinterface.sortingcomponents.peak_pipeline import run_node_pipeline, PeakRetriever, ExtractDenseWaveforms
+from spikeinterface.core.node_pipeline import run_node_pipeline, PeakRetriever, ExtractDenseWaveforms
 from spikeinterface.sortingcomponents.waveforms.neural_network_denoiser import SingleChannelToyDenoiser
