+++ src/spikeinterface/sortingcomponents/tests/test_waveforms/test_waveform_thresholder.py
@@ -4,13 +4,7 @@ import operator
 
 
 from spikeinterface.sortingcomponents.waveforms.waveform_thresholder import WaveformThresholder
-from spikeinterface.sortingcomponents.peak_pipeline import (
-    ExtractDenseWaveforms,
-    ExtractSparseWaveforms,
-    WaveformsNode,
-    PipelineNode,
-    run_peak_pipeline,
-)
+from spikeinterface.core.node_pipeline import ExtractDenseWaveforms, run_peak_pipeline
 
 
 @pytest.fixture(scope="module")
