+++ src/spikeinterface/sortingcomponents/waveforms/neural_network_denoiser.py
@@ -17,7 +17,7 @@ except ImportError:
     HAVE_HUGGINFACE = False
 
 from spikeinterface.core import BaseRecording
-from spikeinterface.sortingcomponents.peak_pipeline import PipelineNode, WaveformsNode, find_parent_of_type
+from spikeinterface.core.node_pipeline import import PipelineNode, WaveformsNode, find_parent_of_type
 from .waveform_utils import to_temporal_representation, from_temporal_representation
