+++ src/spikeinterface/sortingcomponents/waveforms/waveform_thresholder.py
@@ -7,7 +7,7 @@ import operator
 from typing import Literal
 
 from spikeinterface.core import BaseRecording, get_noise_levels
-from spikeinterface.sortingcomponents.peak_pipeline import PipelineNode, WaveformsNode, find_parent_of_type
+from spikeinterface.core.node_pipeline import import PipelineNode, WaveformsNode, find_parent_of_type
 
 
 class WaveformThresholder(WaveformsNode):
